more I32/IV/SSize_t fixes
We are now in a 64bit world and need to get rid of all the wrong
32bit (2GB) size limits.
Some of these fixes seem to be even security relevant, as in the last
2GB series:

chop/chomp of only half of overlarge arrays.

Or ~"a"x2G complement of overlarge strings, silently processing only
the half - as with overlong hash keys.

There was also a smartmatch Array - CodeRef rule, which passed only over
half the array elements.  The Hash part was also wrong, but the wrong number
was not used.

regex match group of >2GB string len.

Missing optimization of inplace substitution via clen overflow.

See GH #123

+++ hv.c
@@ -1677,7 +1677,7 @@ S_hsplit(pTHX_ HV *hv, U32 const oldsize, U32 newsize)
             U32 j = (HeHASH(entry) & newsize);
 #ifdef DEBUGGING
             if (DEBUG_H_TEST_ && DEBUG_v_TEST_) {
-                PerlIO_printf(Perl_debug_log, "HASH split %d->%d\n",i,j);
+                PerlIO_printf(Perl_debug_log, "HASH split %u->%u\n",(unsigned)i,(unsigned)j);
                 deb_hechain(aep[i]);
             }
 #endif
